requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
feedparser>=6.0
tavily-python>=0.5
python-dotenv>=1.0
pydantic>=2.6
aiosqlite>=0.20
playwright>=1.41
